Remove prover
